backend/requirement.txt
fastapi
uvicorn
python-dotenv
google-genai
python-dateutil
numpy
requests
geopy
